// src/common/clients/apis/rest-countries/types.d.ts
export interface Country {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  cioc?: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  gini?: Gini;
  fifa?: string;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  postalCode?: PostalCode;
  borders?: string[];
}

interface PostalCode {
  format: string;
  regex?: string;
}

interface CapitalInfo {
  latlng?: number[];
}

interface CoatOfArms {
  png?: string;
  svg?: string;
}

interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

interface Car {
  signs?: string[];
  side: string;
}

interface Gini {
  '2017'?: number;
  '2013'?: number;
  '2010'?: number;
  '2011'?: number;
  '2018'?: number;
  '2019'?: number;
  '2003'?: number;
  '2012'?: number;
  '2015'?: number;
  '2016'?: number;
  '2009'?: number;
  '1998'?: number;
  '2014'?: number;
  '2006'?: number;
  '2004'?: number;
  '1999'?: number;
  '2005'?: number;
  '2008'?: number;
  '1992'?: number;
}

interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

interface Demonyms {
  eng: Eng;
  fra?: Eng;
}

interface Eng {
  f: string;
  m: string;
}

interface Translations {
  ara: Isl;
  bre: Isl;
  ces: Isl;
  cym: Isl;
  deu: Isl;
  est: Isl;
  fin: Isl;
  fra: Isl;
  hrv?: Isl;
  hun: Isl;
  ita: Isl;
  jpn?: Isl;
  kor: Isl;
  nld: Isl;
  per?: Isl;
  pol: Isl;
  por: Isl;
  rus: Isl;
  slk: Isl;
  spa: Isl;
  srp: Isl;
  swe: Isl;
  tur: Isl;
  urd: Isl;
  zho?: Isl;
}

interface Languages {
  isl?: string;
  jpn?: string;
  fra?: string;
  ara?: string;
  som?: string;
  eng?: string;
  nld?: string;
  pap?: string;
  mri?: string;
  nzs?: string;
  spa?: string;
  sot?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  aym?: string;
  que?: string;
  nau?: string;
  fij?: string;
  hif?: string;
  pau?: string;
  nor?: string;
  cnr?: string;
  est?: string;
  zho?: string;
  msa?: string;
  tam?: string;
  kaz?: string;
  rus?: string;
  crs?: string;
  hye?: string;
  swe?: string;
  cat?: string;
  tur?: string;
  amh?: string;
  grn?: string;
  ces?: string;
  slk?: string;
  fas?: string;
  por?: string;
  pov?: string;
  bar?: string;
  ssw?: string;
  cal?: string;
  cha?: string;
  swa?: string;
  hmo?: string;
  tpi?: string;
  tuk?: string;
  ron?: string;
  mah?: string;
  glv?: string;
  lao?: string;
  afr?: string;
  deu?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  tsn?: string;
  smo?: string;
  nbl?: string;
  nso?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zul?: string;
  kir?: string;
  ton?: string;
  nya?: string;
  khm?: string;
  sqi?: string;
  srp?: string;
  ita?: string;
  sin?: string;
  lav?: string;
  jam?: string;
  pih?: string;
  gsw?: string;
  roh?: string;
  hrv?: string;
  tvl?: string;
  ell?: string;
  ind?: string;
  bos?: string;
  mfe?: string;
  fin?: string;
  arc?: string;
  ckb?: string;
  vie?: string;
  kon?: string;
  lin?: string;
  lua?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  sna?: string;
  toi?: string;
  zib?: string;
  mkd?: string;
  rar?: string;
  tir?: string;
  mya?: string;
  ben?: string;
  gle?: string;
  ltz?: string;
  bis?: string;
  run?: string;
  dan?: string;
  fao?: string;
  pol?: string;
  kor?: string;
  prs?: string;
  pus?: string;
  aze?: string;
  mon?: string;
  uzb?: string;
  fil?: string;
  kat?: string;
  dzo?: string;
  hat?: string;
  ber?: string;
  mey?: string;
  lat?: string;
  sag?: string;
  tha?: string;
  kal?: string;
  ukr?: string;
  heb?: string;
  bul?: string;
  nep?: string;
  nrf?: string;
  zdj?: string;
  hin?: string;
  hun?: string;
  tet?: string;
  mlt?: string;
  nfr?: string;
  bjz?: string;
  gil?: string;
  mlg?: string;
  slv?: string;
  niu?: string;
  urd?: string;
  tkl?: string;
  kin?: string;
  bel?: string;
  div?: string;
  lit?: string;
  tgk?: string;
}

interface Idd {
  root?: string;
  suffixes?: string[];
}

interface Currencies {
  ISK?: ISK;
  JPY?: ISK;
  XPF?: ISK;
  SOS?: ISK;
  JOD?: ISK;
  USD?: ISK;
  DZD?: ISK;
  NZD?: ISK;
  CLP?: ISK;
  EUR?: ISK;
  LSL?: ISK;
  ZAR?: ISK;
  LBP?: ISK;
  NOK?: ISK;
  PEN?: ISK;
  SYP?: ISK;
  AUD?: ISK;
  FJD?: ISK;
  HNL?: ISK;
  XOF?: ISK;
  XCD?: ISK;
  SGD?: ISK;
  KZT?: ISK;
  SCR?: ISK;
  AMD?: ISK;
  CNY?: ISK;
  MYR?: ISK;
  TRY?: ISK;
  ETB?: ISK;
  PYG?: ISK;
  CZK?: ISK;
  IRR?: ISK;
  GHS?: ISK;
  SEK?: ISK;
  GYD?: ISK;
  BND?: ISK;
  SDG?: SDG;
  VES?: ISK;
  SZL?: ISK;
  EGP?: ISK;
  ILS?: ISK;
  STN?: ISK;
  KES?: ISK;
  PGK?: ISK;
  XAF?: ISK;
  TWD?: ISK;
  CVE?: ISK;
  TMT?: ISK;
  MDL?: ISK;
  RON?: ISK;
  GBP?: ISK;
  IMP?: ISK;
  LAK?: ISK;
  NAD?: ISK;
  KWD?: ISK;
  MXN?: ISK;
  WST?: ISK;
  SLL?: ISK;
  KGS?: ISK;
  SHP?: ISK;
  QAR?: ISK;
  TOP?: ISK;
  COP?: ISK;
  MWK?: ISK;
  KYD?: ISK;
  KHR?: ISK;
  RSD?: ISK;
  LKR?: ISK;
  JMD?: ISK;
  CHF?: ISK;
  GNF?: ISK;
  TVD?: ISK;
  IDR?: ISK;
  BAM?: SDG;
  DJF?: ISK;
  MUR?: ISK;
  BWP?: ISK;
  IQD?: ISK;
  VND?: ISK;
  CDF?: ISK;
  NGN?: ISK;
  ZWL?: ISK;
  MKD?: ISK;
  LRD?: ISK;
  CKD?: ISK;
  ARS?: ISK;
  BBD?: ISK;
  ERN?: ISK;
  MMK?: ISK;
  YER?: ISK;
  AWG?: ISK;
  BDT?: ISK;
  SRD?: ISK;
  BRL?: ISK;
  TND?: ISK;
  ANG?: ISK;
  VUV?: ISK;
  CUC?: ISK;
  CUP?: ISK;
  BIF?: ISK;
  BOB?: ISK;
  DKK?: ISK;
  FOK?: ISK;
  PLN?: ISK;
  CAD?: ISK;
  GTQ?: ISK;
  KPW?: ISK;
  UGX?: ISK;
  AFN?: ISK;
  AZN?: ISK;
  MRU?: ISK;
  MNT?: ISK;
  SBD?: ISK;
  UZS?: ISK;
  PHP?: ISK;
  BHD?: ISK;
  ALL?: ISK;
  GEL?: ISK;
  BTN?: ISK;
  INR?: ISK;
  SAR?: ISK;
  BSD?: ISK;
  AED?: ISK;
  HTG?: ISK;
  MAD?: ISK;
  LYD?: ISK;
  CRC?: ISK;
  THB?: ISK;
  TZS?: ISK;
  HKD?: ISK;
  BMD?: ISK;
  UAH?: ISK;
  MOP?: ISK;
  MZN?: ISK;
  DOP?: ISK;
  PAB?: ISK;
  BGN?: ISK;
  SSP?: ISK;
  GMD?: ISK;
  OMR?: ISK;
  RUB?: ISK;
  AOA?: ISK;
  NPR?: ISK;
  JEP?: ISK;
  KMF?: ISK;
  HUF?: ISK;
  FKP?: ISK;
  TTD?: ISK;
  GGP?: ISK;
  GIP?: ISK;
  BZD?: ISK;
  KID?: ISK;
  MGA?: ISK;
  KRW?: ISK;
  UYU?: ISK;
  NIO?: ISK;
  ZMW?: ISK;
  PKR?: ISK;
  RWF?: ISK;
  BYN?: ISK;
  MVR?: ISK;
  TJS?: ISK;
}

interface SDG {
  name: string;
}

interface ISK {
  name: string;
  symbol: string;
}

interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

interface NativeName {
  isl?: Isl;
  jpn?: Isl;
  fra?: Isl;
  ara?: Isl;
  som?: Isl;
  nld?: Isl;
  pap?: Isl;
  eng?: Isl;
  mri?: Isl;
  nzs?: Isl;
  spa?: Isl;
  sot?: Isl;
  nno?: Isl;
  nob?: Isl;
  smi?: Isl;
  aym?: Isl;
  que?: Isl;
  nau?: Isl;
  fij?: Isl;
  hif?: Isl;
  pau?: Isl;
  nor?: Isl;
  cnr?: Isl;
  est?: Isl;
  zho?: Isl;
  msa?: Isl;
  tam?: Isl;
  kaz?: Isl;
  rus?: Isl;
  crs?: Isl;
  hye?: Isl;
  swe?: Isl;
  cat?: Isl;
  tur?: Isl;
  amh?: Isl;
  grn?: Isl;
  ces?: Isl;
  slk?: Isl;
  fas?: Isl;
  por?: Isl;
  pov?: Isl;
  bar?: Isl;
  ssw?: Isl;
  cal?: Isl;
  cha?: Isl;
  swa?: Isl;
  hmo?: Isl;
  tpi?: Isl;
  tuk?: Isl;
  ron?: Isl;
  mah?: Isl;
  glv?: Isl;
  lao?: Isl;
  afr?: Isl;
  deu?: Isl;
  her?: Isl;
  hgm?: Isl;
  kwn?: Isl;
  loz?: Isl;
  ndo?: Isl;
  tsn?: Isl;
  smo?: Isl;
  nbl?: Isl;
  nso?: Isl;
  tso?: Isl;
  ven?: Isl;
  xho?: Isl;
  zul?: Isl;
  kir?: Isl;
  ton?: Isl;
  nya?: Isl;
  khm?: Isl;
  sqi?: Isl;
  srp?: Isl;
  ita?: Isl;
  sin?: Isl;
  lav?: Isl;
  jam?: Isl;
  pih?: Isl;
  gsw?: Isl;
  roh?: Isl;
  hrv?: Isl;
  tvl?: Isl;
  ell?: Isl;
  ind?: Isl;
  bos?: Isl;
  mfe?: Isl;
  fin?: Isl;
  arc?: Isl;
  ckb?: Isl;
  vie?: Isl;
  kon?: Isl;
  lin?: Isl;
  lua?: Isl;
  bwg?: Isl;
  kck?: Isl;
  khi?: Isl;
  ndc?: Isl;
  nde?: Isl;
  sna?: Isl;
  toi?: Isl;
  zib?: Isl;
  mkd?: Isl;
  rar?: Isl;
  tir?: Isl;
  mya?: Isl;
  ben?: Isl;
  gle?: Isl;
  ltz?: Isl;
  bis?: Isl;
  run?: Isl;
  dan?: Isl;
  fao?: Isl;
  pol?: Isl;
  kor?: Isl;
  prs?: Isl;
  pus?: Isl;
  aze?: Isl;
  mon?: Isl;
  uzb?: Isl;
  fil?: Isl;
  kat?: Isl;
  dzo?: Isl;
  hat?: Isl;
  ber?: Isl;
  mey?: Isl;
  lat?: Isl;
  sag?: Isl;
  tha?: Isl;
  kal?: Isl;
  ukr?: Isl;
  heb?: Isl;
  bul?: Isl;
  nep?: Isl;
  nrf?: Isl;
  zdj?: Isl;
  hin?: Isl;
  hun?: Isl;
  tet?: Isl;
  mlt?: Isl;
  nfr?: Isl;
  bjz?: Isl;
  gil?: Isl;
  mlg?: Isl;
  slv?: Isl;
  niu?: Isl;
  urd?: Isl;
  tkl?: Isl;
  kin?: Isl;
  bel?: Isl;
  div?: Isl;
  lit?: Isl;
  tgk?: Isl;
}

interface Isl {
  official: string;
  common: string;
}